pyproject.toml
[tool.poetry]
name = "arachnid-shield-sdk"
version = "0.2.0"
description = "An SDK to work with the Arachnid Shield API"

authors = [
    "Aalekh Patel <aalekh@protectchildren.ca>",
    "Mark Johnston <mark@protectchildren.ca>",
]

readme = "README.md"
license = "LICENSE"

homepage = "https://shield.projectarachnid.com/"
documentation = "https://shield.projectarachnid.com/"
repository = "https://github.com/CdnCentreForChildProtection/arachnid-shield-sdk-python/"
keywords = ["arachnid-shield", "sdk", "arachnid"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

packages = [
    {include = "arachnid_shield_sdk"},
]
include = [
    "CHANGELOG.md",
    "arachnid_shield_sdk/py.typed"
]
exclude = [
    "tests",
    "openapi.json"
]

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-mock = "*"
coverage = "*"
pytest-asyncio = "*"
pytest-httpx = "*"
pytest-cov = "*"
pdqhash = "^0.2.8"
pillow = "^11.3.0"
numpy = "2.2.0"

[tool.poetry.dependencies]
python = "^3.10"
httpx = ">=0.15.4,<0.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py310', 'py311', 'py312', 'py313']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
